build: update and add requirements file for Python 3.9+
Creates minimum and fully working environments for Python 3.9+.

`requirements.txt` produces a minimum working environment without
`gdal`. Postprocessing support is disabled when installing with `pip`
for Python 3.11 as `richdem` is unavailable.

With conda/mamba, `dev_requirements.txt` produces a fully working
environment with `conda install --file dev_requirements.txt`. Installing
with `pip` raises an error as some dependencies are unavailable.

Python 3.11 cannot resolve a dependency issue between the latest
versions of `tornado` and `distributed`, so the maximum `distributed`
version is set to 2023.2. With Python 3.9 and 3.10, `distributed` can be
safely updated to the latest version.

If in doubt, `distributed` versions 2021.3 and 2022.3 should work with
Python 3.9/3.10.